+++ changelog.md
@@ -1,6 +1,7 @@
 # v2.21.0
 
 - Optimized user interface for Windows, and using Smart Notes on lower resolution screens.
+- Upgrade GPT Image model from 1.5 to 2.
 
 # v2.20.0